+++ tests/PHP/BitTorrent/EncoderTest.php
@@ -65,31 +65,29 @@ class EncoderTest extends \PHPUnit_Framework_TestCase {
      *
      * @return array
      */
-    public function getEncodeNumberData() {
+    public function getEncodeIntegerData() {
         return array(
             array(-1, 'i-1e'),
             array(0, 'i0e'),
             array(1, 'i1e'),
-            array(1.0, 'i1e'),
-            array("1", 'i1e'),
-            array("1.0", 'i1e'),
         );
     }
 
     /**
-     * @dataProvider getEncodeNumberData()
-     * @covers PHP\BitTorrent\Encoder::encodeNumber
+     * @dataProvider getEncodeIntegerData()
+     * @covers PHP\BitTorrent\Encoder::encodeInteger
+     * @covers PHP\BitTorrent\Encoder::isInt
      */
-    public function testEncodeNumber($value, $encoded) {
-        $this->assertSame($encoded, $this->encoder->encodeNumber($value));
+    public function testEncodeInteger($value, $encoded) {
+        $this->assertSame($encoded, $this->encoder->encodeInteger($value));
     }
 
     /**
      * @expectedException InvalidArgumentException
-     * @covers PHP\BitTorrent\Encoder::encodeNumber
+     * @covers PHP\BitTorrent\Encoder::encodeInteger
      */
-    public function testEncodeNonNumberAsNumber() {
-        $this->encoder->encodeNumber('one');
+    public function testEncodeNonIntegerAsInteger() {
+        $this->encoder->encodeInteger('one');
     }
 
     /**
@@ -183,11 +181,8 @@ class EncoderTest extends \PHPUnit_Framework_TestCase {
     public function getEncodeData() {
         return array(
             array(1, 'i1e'),
-            array(1.0, 'i1e'),
-            array("1", 'i1e'),
-            array("1.0", 'i1e'),
             array('spam', '4:spam'),
-            array(array(1, 2, 3.0), 'li1ei2ei3ee'),
+            array(array(1, 2, 3), 'li1ei2ei3ee'),
             array(array('foo' => 'bar', 'spam' => 'sucks'), 'd3:foo3:bar4:spam5:suckse'),
         );
     }
@@ -195,6 +190,7 @@ class EncoderTest extends \PHPUnit_Framework_TestCase {
     /**
      * @dataProvider getEncodeData()
      * @covers PHP\BitTorrent\Encoder::encode
+     * @covers PHP\BitTorrent\Encoder::isInt
      */
     public function testEncodeUsingGenericMethod($value, $encoded) {
         $this->assertSame($encoded, $this->encoder->encode($value));